Rozdeleni kodu do modulu
