+++ src/graphics/mod.rs
@@ -1,0 +1,6 @@
+pub mod vertex;
+pub mod depth;
+pub mod pipeline;
+
+pub use vertex::{Vertex, VERTICES, INDICES, as_bytes};
+
